fix(github): use /search/issues for accurate doc estimate
GitHub's /repos/X/issues endpoint switched to cursor pagination and no
longer returns rel="last" in its Link header, so countViaLink falls back
to len(first)=1 — the issues subtotal collapses to ~0 and the dashboard
shows estimates roughly equal to the PR count alone (e.g. 62 estimated
when reality is 62 PRs + 77 issues = 139).

Switch to /search/issues with is:issue / is:pr qualifiers, which returns
total_count directly and splits issues from PRs cleanly. /pulls still
works with the link-header trick but search is consistent and removes
the subtraction that broke the moment one endpoint changed pagination.
