+++ internal/billing/paystack/events.go
@@ -1,13 +1,9 @@
 package paystack
 
 import (
-	"bytes"
 	"encoding/json"
 	"fmt"
-	"strings"
 	"time"
-
-	"github.com/google/uuid"
 
 	"github.com/usehiveloop/hiveloop/internal/billing"
 )
@@ -164,159 +160,3 @@ func (p *Provider) buildState(eventName string, data json.RawMessage) (*billing.
 	return nil, fmt.Errorf("unhandled event shape: %s", eventName)
 }
 
-func (p *Provider) subscriptionState(d subscriptionData) (*billing.SubscriptionState, error) {
-	orgID, err := extractOrgID(d.Customer.Metadata)
-	if err != nil {
-		return nil, err
-	}
-	state := &billing.SubscriptionState{
-		ExternalSubscriptionID: d.SubscriptionCode,
-		ExternalCustomerID:     d.Customer.CustomerCode,
-		OrgID:                  orgID,
-		PlanSlug:               p.cfg.Plans.SlugForPlanCode(d.Plan.PlanCode),
-		Status:                 mapSubscriptionStatus(d.Status),
-	}
-	if d.CreatedAt != nil {
-		state.CurrentPeriodStart = *d.CreatedAt
-	}
-	if d.NextPaymentDate != nil {
-		state.CurrentPeriodEnd = *d.NextPaymentDate
-	}
-	return state, nil
-}
-
-func (p *Provider) chargeState(d chargeData, planCode string) (*billing.SubscriptionState, error) {
-	orgID, err := extractOrgID(d.Customer.Metadata)
-	if err != nil {
-		return nil, err
-	}
-	if orgID == uuid.Nil {
-		// Customer.metadata is often null on charge events. Fall back to the
-		// transaction-level metadata, which carries the {org_id, plan_slug}
-		// we set on /transaction/initialize.
-		orgID, err = extractOrgIDFromRaw(d.Metadata)
-		if err != nil {
-			return nil, err
-		}
-	}
-	planSlug := p.cfg.Plans.SlugForPlanCode(planCode)
-	if planSlug == "" {
-		// Charge for a plan_code we don't manage — someone else's plan on
-		// the same Paystack account, or a dashboard-created one. Ignore.
-		return nil, nil
-	}
-	state := &billing.SubscriptionState{
-		ExternalCustomerID: d.Customer.CustomerCode,
-		OrgID:              orgID,
-		PlanSlug:           planSlug,
-		Status:             billing.StatusActive,
-		ChargeReference:    d.Reference,
-		ChargeAmount:       d.Amount,
-		ChargedAt:          d.PaidAt,
-		CardLast4:          d.Authorization.Last4,
-		CardBrand:          d.Authorization.Brand,
-		CardExpMonth:       d.Authorization.ExpMonth,
-		CardExpYear:        d.Authorization.ExpYear,
-		AuthorizationCode:  d.Authorization.AuthorizationCode,
-	}
-	if d.PaidAt != nil {
-		state.CurrentPeriodStart = *d.PaidAt
-	}
-	return state, nil
-}
-
-// invoiceState builds state for invoice.* events, where the customer is a
-// sibling of subscription at the data root and no plan object is present.
-// PlanSlug is left blank — the handler resolves the plan via the existing
-// subscription row keyed by subscription_code.
-func (p *Provider) invoiceState(sub subscriptionData, customer customerPayload) (*billing.SubscriptionState, error) {
-	orgID, err := extractOrgID(customer.Metadata)
-	if err != nil {
-		return nil, err
-	}
-	state := &billing.SubscriptionState{
-		ExternalSubscriptionID: sub.SubscriptionCode,
-		ExternalCustomerID:     customer.CustomerCode,
-		OrgID:                  orgID,
-		Status:                 mapSubscriptionStatus(sub.Status),
-	}
-	if sub.NextPaymentDate != nil {
-		state.CurrentPeriodEnd = *sub.NextPaymentDate
-	}
-	return state, nil
-}
-
-// parseChargePlanCode interprets Paystack's variable shape for `data.plan`
-// on charge.success events. Observed shapes in the wild (all real):
-//
-//	(missing) | null | {}      → empty (one-off charge, not a subscription)
-//	"PLN_abc"                  → plan_code string directly (older format)
-//	{"plan_code":"PLN_abc",…}  → plan_code inside an object (newer format)
-//
-// Returns the empty string when no usable plan_code can be extracted.
-func parseChargePlanCode(raw json.RawMessage) string {
-	trimmed := bytes.TrimSpace(raw)
-	if len(trimmed) == 0 || bytes.Equal(trimmed, []byte("null")) || bytes.Equal(trimmed, []byte("{}")) {
-		return ""
-	}
-	var asString string
-	if err := json.Unmarshal(raw, &asString); err == nil {
-		return strings.TrimSpace(asString)
-	}
-	var asObj planPayload
-	if err := json.Unmarshal(raw, &asObj); err == nil {
-		return asObj.PlanCode
-	}
-	return ""
-}
-
-// extractOrgID pulls the org uuid from customer metadata. EnsureCustomer
-// always sets {org_id, org_name} when we create the customer, so every
-// event tied to an API-created customer will have this populated. Returns
-// (uuid.Nil, nil) when absent — the webhook handler's resolveOrgID falls
-// back to looking up the org by subscription_code or customer_code. An
-// invalid (non-UUID) org_id is a hard error.
-func extractOrgID(meta map[string]string) (uuid.UUID, error) {
-	raw, ok := meta["org_id"]
-	if !ok || raw == "" {
-		return uuid.Nil, nil
-	}
-	parsed, err := uuid.Parse(raw)
-	if err != nil {
-		return uuid.Nil, fmt.Errorf("invalid org_id %q: %w", raw, err)
-	}
-	return parsed, nil
-}
-
-// extractOrgIDFromRaw decodes a raw JSON value into a string map and pulls
-// out org_id. Paystack's transaction-level `metadata` can arrive as
-// {"org_id":"…"}, an empty object, null, or even a number — we silently
-// give up on anything that doesn't parse as a string map.
-func extractOrgIDFromRaw(raw json.RawMessage) (uuid.UUID, error) {
-	trimmed := bytes.TrimSpace(raw)
-	if len(trimmed) == 0 || bytes.Equal(trimmed, []byte("null")) {
-		return uuid.Nil, nil
-	}
-	var meta map[string]string
-	if err := json.Unmarshal(raw, &meta); err != nil {
-		return uuid.Nil, nil
-	}
-	return extractOrgID(meta)
-}
-
-// mapSubscriptionStatus normalises Paystack's status strings onto our own
-// lifecycle. Paystack uses "non-renewing" while the subscription is
-// winding down, and either "complete" (billing cycles exhausted) or
-// "cancelled" (user-cancelled at period end) once the subscription is
-// fully terminated.
-func mapSubscriptionStatus(paystackStatus string) billing.SubscriptionStatus {
-	switch strings.ToLower(paystackStatus) {
-	case "active", "attention":
-		return billing.StatusActive
-	case "non-renewing", "cancelled", "canceled":
-		return billing.StatusCanceled
-	case "complete", "completed":
-		return billing.StatusRevoked
-	}
-	return billing.StatusActive
-}
