+++ internal/rag/connectors/github/estimate.go
@@ -5,6 +5,7 @@ import (
 	"fmt"
 	"net/http"
 	"net/url"
+	"strings"
 
 	"github.com/usehiveloop/hiveloop/internal/rag/connectors/interfaces"
 )
@@ -21,63 +22,48 @@ func (c *GithubConnector) EstimateTotal(ctx context.Context, src interfaces.Sour
 	total := 0
 	for _, repo := range cfg.Repositories {
 		fullName := cfg.RepoOwner + "/" + repo
-		var prCount int
-		if cfg.IncludePRs || cfg.IncludeIssues {
-			n, err := c.client.countViaLink(ctx, "/repos/"+fullName+"/pulls", cfg.StateFilter)
-			if err != nil {
-				return 0, fmt.Errorf("github estimate: pulls %s: %w", fullName, err)
-			}
-			prCount = n
-		}
 		if cfg.IncludePRs {
-			total += prCount
+			n, err := c.client.searchCount(ctx, fullName, "pr", cfg.StateFilter)
+			if err != nil {
+				return 0, fmt.Errorf("github estimate: prs %s: %w", fullName, err)
+			}
+			total += n
 		}
 		if cfg.IncludeIssues {
-			// /issues returns issues + PRs; subtract PRs to get the
-			// actual issue count the connector will emit (it filters
-			// PR-shaped items out of the issues stream).
-			n, err := c.client.countViaLink(ctx, "/repos/"+fullName+"/issues", cfg.StateFilter)
+			n, err := c.client.searchCount(ctx, fullName, "issue", cfg.StateFilter)
 			if err != nil {
 				return 0, fmt.Errorf("github estimate: issues %s: %w", fullName, err)
 			}
-			issuesOnly := n - prCount
-			if issuesOnly < 0 {
-				issuesOnly = 0
-			}
-			total += issuesOnly
+			total += n
 		}
 	}
 	return total, nil
 }
 
-// countViaLink probes the list endpoint at per_page=1 and reads the
-// rel="last" page number from the Link header — GitHub omits Link for
-// empty/single-item lists, so fall back to len(body).
-func (c *Client) countViaLink(ctx context.Context, path, state string) (int, error) {
-	q := url.Values{}
-	q.Set("per_page", "1")
-	q.Set("page", "1")
-	if state != "" {
-		q.Set("state", state)
-	}
+type searchIssuesResponse struct {
+	TotalCount int `json:"total_count"`
+}
 
-	var hdr http.Header
-	var first []map[string]any
+func (c *Client) searchCount(ctx context.Context, fullName, kind, state string) (int, error) {
+	parts := []string{"repo:" + fullName, "is:" + kind}
+	if state != "" && state != "all" {
+		parts = append(parts, "state:"+state)
+	}
+	q := url.Values{}
+	q.Set("q", strings.Join(parts, " "))
+	q.Set("per_page", "1")
+
+	var resp searchIssuesResponse
 	err := withRateLimitRetry(ctx, func() error {
-		got, h, err := getJSON[[]map[string]any](ctx, c.p, http.MethodGet, path, q)
+		got, _, err := getJSON[searchIssuesResponse](ctx, c.p, http.MethodGet, "/search/issues", q)
 		if err != nil {
 			return err
 		}
-		first = got
-		hdr = h
+		resp = got
 		return nil
 	})
 	if err != nil {
 		return 0, err
 	}
-	if last, ok := lastPageNumber(hdr); ok {
-		return last, nil
-	}
-	return len(first), nil
+	return resp.TotalCount, nil
 }
-
